Revert "refactor: introduce Container - Presenter"
