+++ src/components/Slotmachine/Slot/index.ts
@@ -1,4 +1,2 @@
 export { useSlot } from "./hooks";
 export { Slot } from "./Slot";
-export type { SlotProps } from "./Slot";
-export { SlotContainer } from "./Container";
