--- src/components/Slotmachine/index.ts
@@ -1,8 +1,0 @@
-export type {
-  IncompleteSlotItem,
-  IncompleteSlots,
-  StoppedSlotItem,
-  StoppedSlots,
-  Slots,
-} from "./types";
-export { Slotmachine } from "./Slotmachine";
