+++ src/components/Slotmachine/index.tsx
@@ -1,20 +1,22 @@
 import React, { useCallback, useRef, useState } from "react";
-import { Slots, StoppedSlots } from "./types";
-import { SlotmachinePresenterProps } from "./Presenter";
-import { useSlot } from "./Slot";
+import { Slot, useSlot } from "./Slot";
 
-export type SlotmachineContainerProps = Readonly<{
+export type IncompleteSlotItem = string | undefined;
+export type IncompleteSlots = Array<IncompleteSlotItem>;
+export type StoppedSlotItem = string;
+export type StoppedSlots = Array<StoppedSlotItem>;
+export type Slots = StoppedSlots | IncompleteSlots;
+
+type SlotmachineProps = Readonly<{
   symbols: readonly string[];
   expected: readonly string[];
   onStopped: (result: StoppedSlots) => void;
-  renderPresenter: (props: SlotmachinePresenterProps) => React.ReactElement;
 }>;
 
-export const SlotmachineContainer: React.FC<SlotmachineContainerProps> = ({
+export const Slotmachine: React.FC<SlotmachineProps> = ({
   symbols: _symbols,
   expected,
   onStopped,
-  renderPresenter,
 }) => {
   // explicitly *freeze* symbols with initial value
   const [symbols] = useState(_symbols);
@@ -63,10 +65,29 @@ export const SlotmachineContainer: React.FC<SlotmachineContainerProps> = ({
     setMachineState("rolling");
     propsAndStarts.forEach(({ start }) => start());
   };
-  const propss = propsAndStarts.map(({ props }) => props);
-  return renderPresenter({
-    propss,
-    start,
-    isRolling: machineState == "rolling",
-  });
+  return (
+    <div>
+      <table>
+        <tbody>
+          <tr>
+            {propsAndStarts.map(({ props }) => (
+              <td key={props.initialIndex}>
+                <Slot {...props} />
+              </td>
+            ))}
+          </tr>
+          <tr>
+            <td rowSpan={symbols.length}>
+              <button
+                onClick={() => start()}
+                disabled={machineState === "rolling"}
+              >
+                start!!
+              </button>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  );
 };
