--- src/components/Slotmachine/types.ts
@@ -1,5 +1,0 @@
-export type IncompleteSlotItem = string | undefined;
-export type IncompleteSlots = Array<IncompleteSlotItem>;
-export type StoppedSlotItem = string;
-export type StoppedSlots = Array<StoppedSlotItem>;
-export type Slots = StoppedSlots | IncompleteSlots;
